# Conflicts:
#	lucene/common-build.xml
#	lucene/ivy-versions.properties
#	solr/bin/solr
#	solr/build.xml
#	solr/contrib/analysis-extras/build.xml
#	solr/core/src/java/org/apache/solr/util/SolrCLI.java
#	solr/server/resources/log4j.properties
#	solr/server/scripts/cloud-scripts/log4j.properties
#	solr/solrj/src/java/org/apache/solr/common/params/CommonParams.java
#	solr/webapp/web/index.html
#	solr/webapp/web/js/angular/app.js

+++ lucene/misc/src/test/org/apache/lucene/misc/SweetSpotSimilarityTest.java
@@ -55,7 +55,7 @@ public class SweetSpotSimilarityTest extends LuceneTestCase {
     dir.close();
     Explanation norm = findExplanation(expl, "fieldNorm");
     assertNotNull(norm);
-    return norm.getValue();
+    return norm.getValue().floatValue();
   }
 
   private static Explanation findExplanation(Explanation expl, String text) {
